+++ src/components/ControlsContent/ControlsContent.tsx
@@ -1,8 +1,8 @@
 import Button from "components/common/Button";
-import NotesPanel from "../NotesPanel";
+import NotesPanel from "components/NotesPanel";
 import React from "react";
-import ScoreControls from "../ScoreControls";
-import Settings from "components/common/Settings";
+import ScoreControls from "components/ScoreControls";
+import Settings from "components/Settings";
 import About from "components/About";
 import VersionNotes from "components/Changelog";
 import useStyles from "./ControlsContent.styles";
