+++ src/components/ScoreControls/Score/Score.styles.ts
