+++ src/components/ScoreControls/Score/Score.tsx
