+++ src/components/ScoreControls/Score/index.ts
