+++ src/components/ScoresPanel/CircularProgressBarScores/CircularProgressBar/CircularProgressBar.styles.ts
