+++ src/components/ScoresPanel/CircularProgressBarScores/CircularProgressBar/CircularProgressBar.tsx
