+++ src/components/ScoresPanel/CircularProgressBarScores/CircularProgressBar/index.ts
