--- src/components/common/ProgressBar/index.ts
@@ -1,3 +1,0 @@
-import ProgressBar from "./ProgressBar";
-
-export default ProgressBar;
